// src/app/models/calculator-invoices.ts
export class CalculatorInvoice {
  id: number;
  name: string;
  address: string;
  document: string;
  phone: string;
  status: string;
  surname: string;
  documentType: string;
  email: string;
  reference: string;
  createdAt: string;
  expirationDate: string;
  days: string;
  valueAddedTax: number;
  subtotal: number;
  total: number;
  saldoDocumentoAdicional: number;
  currency: string;
  urlDoc: string;
  nit: string;
  percentage: number;
  calculatedValue: number;

  constructor() { }
}